Move base fragment and utils to core module
