+++ app/src/main/java/com/example/cleannews/api/NewsApi.kt
@@ -1,7 +1,7 @@
 package com.example.cleannews.api
 
 import com.example.cleannews.model.NewsResponse
-import com.example.cleannews.util.Constants.Companion.API_KEY
+import com.example.core.util.Constants.Companion.API_KEY
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
