+++ app/src/main/java/com/example/cleannews/ui/breakingnews/BreakingNewsFragment.kt
@@ -2,26 +2,20 @@ package com.example.cleannews.ui.breakingnews
 
 import android.os.Bundle
 import android.util.Log
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.widget.AbsListView
-import android.widget.TextView
 import android.widget.Toast
-import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.cleannews.NewsActivity
 import com.example.cleannews.NewsViewModel
 import com.example.cleannews.R
 import com.example.cleannews.adapter.NewsAdapter
 import com.example.cleannews.databinding.FragmentBreakingNewsBinding
-import com.example.cleannews.ui.BaseFragment
-import com.example.cleannews.util.Constants.Companion.QUERY_PAGE_SIZE
-import com.example.cleannews.util.Resource
+import com.example.core.BaseFragment
+import com.example.core.util.Constants.Companion.QUERY_PAGE_SIZE
+import com.example.core.util.Resource
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.android.synthetic.main.fragment_breaking_news.*
